refinement for bytes to BGR IMG
